pyproject.toml
[tool.poetry]
name = "docker-workshop"
version = "0.1.0"
description = "A simple Flask web server for a Docker workshop"
authors = []
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.1"
psycopg2-binary = "^2.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"

[tool.poetry.scripts]
dev = "app.main:run"
test = "pytest:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
